pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyaccelerate"
dynamic = ["version"]
description = "High-performance Python acceleration engine — CPU, threads, virtual threads, multi-GPU and virtualization."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "Guilherme Pinheiro", email = "guilherme@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Hardware",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
keywords = ["acceleration", "gpu", "cuda", "opencl", "threads", "performance", "optimization"]
dependencies = [
    "psutil>=5.9",
]

[project.optional-dependencies]
cuda = ["cupy-cuda12x>=13.0"]
opencl = ["pyopencl>=2024.1"]
intel = ["dpctl>=0.16", "dpnp>=0.14"]
all-gpu = ["cupy-cuda12x>=13.0", "pyopencl>=2024.1"]
numpy = ["numpy>=1.26"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "ruff>=0.4",
    "pre-commit>=3.7",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

[project.urls]
Homepage = "https://github.com/GuilhermeP96/pyaccelerate"
Repository = "https://github.com/GuilhermeP96/pyaccelerate"
Issues = "https://github.com/GuilhermeP96/pyaccelerate/issues"

[project.scripts]
pyaccelerate = "pyaccelerate.cli:main"

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pyaccelerate = ["py.typed"]

# ── Ruff ──────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["pyaccelerate"]

# ── Pytest ────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "gpu: requires GPU compute backend",
    "slow: long-running benchmark tests",
]

# ── Mypy ──────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["cupy.*", "pyopencl.*", "dpctl.*", "dpnp.*", "psutil.*"]
ignore_missing_imports = true

# ── Coverage ──────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src/pyaccelerate"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
